refactor: update the lib folder structures
