+++ src/components/RedeemUI/ReimburseStatusUI/index.tsx
@@ -16,9 +16,9 @@ import InterlayDenimOrKintsugiMidnightOutlinedButton from '@/components/buttons/
 import ErrorFallback from '@/components/ErrorFallback';
 import PrimaryColorSpan from '@/components/PrimaryColorSpan';
 import { RELAY_CHAIN_NATIVE_TOKEN, RELAY_CHAIN_NATIVE_TOKEN_SYMBOL, WRAPPED_TOKEN_SYMBOL } from '@/config/relay-chains';
+import { useSubstrateSecureState } from '@/lib/substrate/substrate-context';
 import RequestWrapper from '@/pages/Bridge/RequestWrapper';
 import { REDEEMS_FETCHER } from '@/services/fetchers/redeems-fetcher';
-import { useSubstrateSecureState } from '@/substrate-lib/substrate-context';
 import { KUSAMA, POLKADOT } from '@/utils/constants/relay-chain-names';
 import { getColorShade } from '@/utils/helpers/colors';
 import { getTokenPrice } from '@/utils/helpers/prices';
