+++ src/lib/substrate/substrate-context.tsx
@@ -15,7 +15,7 @@ import { useLocalStorage } from 'react-use';
 import { APP_NAME } from '@/config/relay-chains';
 import { SELECTED_ACCOUNT_LOCAL_STORAGE_KEY } from '@/config/wallets';
 
-import * as constants from '../constants';
+import * as constants from '../../constants';
 
 const UNABLE_TO_INITIALISE_OPTIONS_MORE_THAN_ONCE = 'Unable to initialise options more than once';
