+++ src/pages/Bridge/BurnForm/index.tsx
@@ -31,7 +31,7 @@ import {
   WrappedTokenLogoIcon
 } from '@/config/relay-chains';
 import { BALANCE_MAX_INTEGER_LENGTH } from '@/constants';
-import { useSubstrateSecureState } from '@/substrate-lib/substrate-context';
+import { useSubstrateSecureState } from '@/lib/substrate/substrate-context';
 import { ForeignAssetIdLiteral } from '@/types/currency';
 import { KUSAMA, POLKADOT } from '@/utils/constants/relay-chain-names';
 import STATUSES from '@/utils/constants/statuses';
