+++ src/pages/Bridge/RedeemForm/index.tsx
@@ -36,8 +36,8 @@ import {
   WRAPPED_TOKEN_SYMBOL
 } from '@/config/relay-chains';
 import { BALANCE_MAX_INTEGER_LENGTH, BTC_ADDRESS_REGEX } from '@/constants';
+import { useSubstrateSecureState } from '@/lib/substrate/substrate-context';
 import ParachainStatusInfo from '@/pages/Bridge/ParachainStatusInfo';
-import { useSubstrateSecureState } from '@/substrate-lib/substrate-context';
 import { ForeignAssetIdLiteral } from '@/types/currency';
 import { KUSAMA, POLKADOT } from '@/utils/constants/relay-chain-names';
 import STATUSES from '@/utils/constants/statuses';
