+++ src/pages/Transfer/TransferForm/index.tsx
@@ -17,7 +17,7 @@ import SubmitButton from '@/components/SubmitButton';
 import TextField from '@/components/TextField';
 import Tokens, { TokenOption } from '@/components/Tokens';
 import InterlayButtonBase from '@/components/UI/InterlayButtonBase';
-import { useSubstrateSecureState } from '@/substrate-lib/substrate-context';
+import { useSubstrateSecureState } from '@/lib/substrate/substrate-context';
 import { KUSAMA, POLKADOT } from '@/utils/constants/relay-chain-names';
 import STATUSES from '@/utils/constants/statuses';
 import isValidPolkadotAddress from '@/utils/helpers/is-valid-polkadot-address';
