+++ src/pages/Vaults/Vault/RequestIssueModal/index.tsx
@@ -28,9 +28,9 @@ import {
   WRAPPED_TOKEN_SYMBOL,
   WrappedTokenLogoIcon
 } from '@/config/relay-chains';
+import { useSubstrateSecureState } from '@/lib/substrate/substrate-context';
 import SubmittedIssueRequestModal from '@/pages/Bridge/IssueForm/SubmittedIssueRequestModal';
 import { useGovernanceTokenBalance } from '@/services/hooks/use-token-balance';
-import { useSubstrateSecureState } from '@/substrate-lib/substrate-context';
 import { ForeignAssetIdLiteral } from '@/types/currency';
 import { KUSAMA, POLKADOT } from '@/utils/constants/relay-chain-names';
 import STATUSES from '@/utils/constants/statuses';
