+++ src/pages/Vaults/Vault/VaultDashboard.tsx
@@ -8,8 +8,8 @@ import { Stack } from '@/component-library';
 import { ProgressCircle } from '@/component-library/ProgressCircle';
 import ErrorFallback from '@/components/ErrorFallback';
 import PrimaryColorEllipsisLoader from '@/components/PrimaryColorEllipsisLoader';
+import { useSubstrateSecureState } from '@/lib/substrate/substrate-context';
 import MainContainer from '@/parts/MainContainer';
-import { useSubstrateSecureState } from '@/substrate-lib/substrate-context';
 import { URL_PARAMETERS } from '@/utils/constants/links';
 import { useGetCurrencies } from '@/utils/hooks/api/use-get-currencies';
 import { useGetVaultData } from '@/utils/hooks/api/vaults/use-get-vault-data';
