+++ src/parts/Topbar/index.tsx
@@ -14,8 +14,8 @@ import Tokens from '@/components/Tokens';
 import InterlayLink from '@/components/UI/InterlayLink';
 import { ACCOUNT_ID_TYPE_NAME } from '@/config/general';
 import { GOVERNANCE_TOKEN, GOVERNANCE_TOKEN_SYMBOL } from '@/config/relay-chains';
+import { useSubstrateSecureState } from '@/lib/substrate/substrate-context';
 import AccountModal from '@/parts/AccountModal';
-import { useSubstrateSecureState } from '@/substrate-lib/substrate-context';
 import { BitcoinNetwork } from '@/types/bitcoin';
 
 import GetGovernanceTokenUI from './GetGovernanceTokenUI';
