+++ src/utils/hooks/use-account-id.ts
@@ -4,7 +4,7 @@ import * as React from 'react';
 import { useSelector } from 'react-redux';
 
 import { StoreType } from '@/common/types/util.types';
-import { useSubstrateSecureState } from '@/substrate-lib/substrate-context';
+import { useSubstrateSecureState } from '@/lib/substrate/substrate-context';
 
 const useAccountId = (accountAddress?: string): AccountId | undefined => {
   const { selectedAccount } = useSubstrateSecureState();
